requirements.txt
streamlit
Pillow
PyMuPDF
pdf2image
